MULE-14701: Connectors doesn't communicate correctly if their POJOs properties are nulleable or not
